Welcome Screen | extract dedicated modules
