+++ modules/project/ui/resources/META-INF/sap.commerce.toolset-project-ui.xml
@@ -33,21 +33,4 @@
                              provider="sap.commerce.toolset.project.options.ProjectSettingsConfigurableProvider"/>
 
     </extensions>
-
-    <actions>
-        <action id="sap.developers.toolset.removeFromRecentProjects"
-                class="sap.commerce.toolset.project.welcomescreen.actionSystem.RemoveSapCommerceProjectAction"
-                text="hybris.welcometab.button.remove.from.recent.projects">
-            <keyboard-shortcut keymap="$default" first-keystroke="DELETE"/>
-        </action>
-
-        <group id="sap.developers.toolset.welcomeTab.projectContextMenu">
-            <reference ref="sap.developers.toolset.removeFromRecentProjects"/>
-        </group>
-    </actions>
-
-    <extensions defaultExtensionNs="com.intellij">
-        <welcomeTabFactory id="SapCommerceCloudWelcomeTab" order="first"
-                           implementation="sap.commerce.toolset.project.welcomescreen.ui.SapCommerceWelcomeTabFactory"/>
-    </extensions>
 </idea-plugin>
