+++ modules/welcomescreen/core/src/sap/commerce/toolset/welcomescreen/presentation/RecentSapCommerceProjectGitBranch.kt
@@ -16,16 +16,9 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.project.welcomescreen.ui
+package sap.commerce.toolset.welcomescreen.presentation
 
-import com.intellij.openapi.Disposable
-import com.intellij.openapi.wm.WelcomeScreen
-import com.intellij.openapi.wm.WelcomeScreenTab
-import com.intellij.openapi.wm.WelcomeTabFactory
-
-class SapCommerceWelcomeTabFactory : WelcomeTabFactory {
-    override fun createWelcomeTabs(
-        ws: WelcomeScreen,
-        parentDisposable: Disposable
-    ): MutableList<WelcomeScreenTab> = mutableListOf(SapCommerceWelcomeTab(parentDisposable))
+sealed interface RecentSapCommerceProjectGitBranch {
+    data class Named(val name: String) : RecentSapCommerceProjectGitBranch
+    data object NotAGitRepo : RecentSapCommerceProjectGitBranch
 }
