+++ modules/welcomescreen/core/src/sap/commerce/toolset/welcomescreen/reader/HybrisProjectSettingsReader.kt
@@ -16,12 +16,12 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.project.welcomescreen.reader
+package sap.commerce.toolset.welcomescreen.reader
 
 import sap.commerce.toolset.HybrisConstants
 import sap.commerce.toolset.project.ProjectConstants
-import sap.commerce.toolset.project.welcomescreen.presentation.RecentSapCommerceProjectSettings
 import sap.commerce.toolset.util.fileExists
+import sap.commerce.toolset.welcomescreen.presentation.RecentSapCommerceProjectSettings
 import java.nio.file.Files
 import java.nio.file.Path
