+++ modules/welcomescreen/ui/build.gradle.kts
@@ -45,6 +45,7 @@ dependencies {
     implementation(project(":meta-core"))
     implementation(project(":project-core"))
     implementation(project(":project-extensioninfo"))
+    implementation(project(":welcomescreen-core"))
 
     intellijPlatform {
         intellijIdea(properties("intellij.version")) {
