+++ modules/welcomescreen/ui/src/sap/commerce/toolset/welcomescreen/ui/SapCommerceWelcomeTabFactory.kt
@@ -16,9 +16,16 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.project.welcomescreen.presentation
+package sap.commerce.toolset.welcomescreen.ui
 
-sealed interface RecentSapCommerceProjectGitBranch {
-    data class Named(val name: String) : RecentSapCommerceProjectGitBranch
-    data object NotAGitRepo : RecentSapCommerceProjectGitBranch
+import com.intellij.openapi.Disposable
+import com.intellij.openapi.wm.WelcomeScreen
+import com.intellij.openapi.wm.WelcomeScreenTab
+import com.intellij.openapi.wm.WelcomeTabFactory
+
+class SapCommerceWelcomeTabFactory : WelcomeTabFactory {
+    override fun createWelcomeTabs(
+        ws: WelcomeScreen,
+        parentDisposable: Disposable
+    ): MutableList<WelcomeScreenTab> = mutableListOf(SapCommerceWelcomeTab(parentDisposable))
 }
